Move custom LinkRenderer out of PaginationHelper
Works around `Rails/HelperInstanceVariable` offense false positives.
